Ajout de la génération de playlist ranked beatsaber

+++ ApolloBot.BeatSaver/Models/BeatSaverSongMetadataApi.cs
@@ -30,6 +30,6 @@ namespace ApolloBot.BeatSaver.Models
         public string AuthorName { get; set; }
 
         [DataMember(Name = "bpm")]
-        public int Bpm { get; set; }
+        public float Bpm { get; set; }
     }
 }
